created a constants package that aims to manage all the constants that are used in several other packages. also renamed realbus to snesbus

+++ memory/test_bus.go
@@ -33,6 +33,6 @@ func (b *TestBus) GetOpenBus() byte {
 	return 0
 }
 
-func (b *TestBus) GetAccessClass(address uint32) byte {
+func (b *TestBus) GetAccessClass(address uint32) uint64 {
 	return FAST_REGION
 }
